MLK-12340: CAAM: system hang during cryptomgr_test
During the crypto manager self test, a aead encryption opertation
is carried out. This operations allocates a aead request, which is
handed to the CAAM driver. The CAAM driver allocates and maps the
required structures.

During the allocation aead extended descriptor, a DMA to device
mapping and synchronization are required. The order of this two
operations matter, which should be map and then sync. Otherwise,
there will be NULL pointer exception.

This patch fix the order of this two operations, from sync-then-map
to map-then-sync.

Signed-off-by: Ulises Cardenas <raul.casas@nxp.com>