MGS-2513 [#imx-320] [imx7ulp] RTC suspend/resume stress test met resume failed.
When on condition that suspend/resume repeatly called and no other gpu command, the command buffer will filled with command comes from gckCOMMAND_Start and not gckCOMMAND_Stall to the hardware, thus lead to command buffer full and waiting for a hardware signal. Because these commands not submit to the hardware to execute, hardware won't  emit such signal, make the gpu hang.

Date: Jan 6, 2017
Signed-off-by: Yuchou Gan <yuchou.gan@nxp.com>