MLK-14254: ASoC: imx_pcm_rpmsg: fix cmd dropped by work queue
The test case is to playback a bitstream, then repeat ctrl+z and fg,
several times later, the playback is failed to continue.

The reason is if the work is pending in work queue, send second time
of this work, the second work is dropped by work queue. so use one
work for one cmd is not fit for audio case. use a work loop for audio
cmd to fix this issue.

Signed-off-by: Shengjiu Wang <shengjiu.wang@freescale.com>
(cherry picked from commit ca3a61be1b1c80791dd90d3cabc9b73ceb38b24c)