MLK-14301: imx dts: Fix cpu regulator references
On imx_4.9.y we no longer override the cpu arm/soc/pu-supply properties
to link to pmic directly but rather rely on the regulator core
propagating a set_voltage on a bypassed regulator upwards.

This change was already performance for imx6qdl, now do it for all the
boards using ldo-bypass.

Signed-off-by: Leonard Crestez <leonard.crestez@nxp.com>