MLK-14286-7 ARM: clk: imx7ulp: mpll can only be used by mipi dsi
The mpll can only be configured by mipi dphy and it is disabled
by default. And it is also not a fixed frequency clock source.
So replace it with a dummy clock for 'periph_slow_sels' to avoid
any peripheral except mipi to use it for clock source.

Signed-off-by: Fancy Fang <chen.fang@nxp.com>
(cherry picked from commit 80c52ca0f11fa23bd036990b45127725262d5ae1)